[parser] split implementation in modules
