+++ parser/src/scanner.rs
@@ -1,0 +1,13 @@
+pub trait Scanner {
+    /// peek the next character without consuming it
+    fn peek(&self) -> Option<char>;
+
+    /// consume and return the next character
+    fn next(&mut self) -> Option<char>;
+
+    /// make a checkpoint of the current position in the input (for backtracking)
+    fn checkpoint(&self) -> usize;
+
+    /// restore the input to a previous position (backtracking)
+    fn rollback(&mut self, pos: usize);
+}
